2nd Revision Exercise 5-3 = Fixed logic
Fixed logic.
Previously, i fixed the displayed purchased items list where the list will only display list of purchased clothing based on customer's body measurement, and matched the value to the list of available clothing with the same size.

But, the problem is the total cost that the customer need to pay is still false. It's still sum all the available clothing to purchase even though the customer only purchased one item/two items. So i fixed it and now, the total cost will be showed the correct amount of money the customer need to pay.

And also this commit marked the final revision for Exercise 5-3.
Sorry for my english, still practicing my writing. I hope my english improve each time i wrote more in english. More to come.